pyproject.toml
[project]
name = "wordpress-mcp-server"
version = "0.1.0"
description = "MCP server for WordPress site management with SEO capabilities"
authors = [{name = "Shawn Shirazi"}]
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "requests>=2.31.0",
    "paramiko>=3.4.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "Pillow>=10.0.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
